# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PruvodceStudenta UTB - Api Documentation",
        "description": "Api Documentation for UTB's PruvodceStudenta",
        "contact": {
            "name": "Sedlar David",
            "email": "sedlar@sutb.cz"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/api/pages/landing/search": {
            "post": {
                "tags": [
                    "Pages"
                ],
                "summary": "Search",
                "operationId": "d2bf13a5d96ada6971a699b5457b226c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "query": {
                                        "type": "string"
                                    },
                                    "locations": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "sections": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "query": "hospoda",
                                    "locations": [
                                        1
                                    ],
                                    "sections": [
                                        2,
                                        4
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
...
}